Add checking of thresholds for send data immediately
